+++ conans/test/unittests/client/cmd/export_test.py
@@ -3,11 +3,8 @@ import shutil
 import unittest
 from collections import namedtuple
 
-import pytest
-
 from conans.client.cmd.export import _replace_scm_data_in_conanfile
 from conans.client.loader import _parse_conanfile
-from conans.model.ref import ConanFileReference
 from conans.model.scm import SCMData
 from conans.test.utils.test_files import temp_folder
 from conans.test.utils.tools import TestClient, TestServer, TurboTestClient
@@ -120,108 +117,3 @@ class ConanLib(ConanFile):
         after_scm = ''
         after_recipe = ''
         self._do_actual_test(scm_data=scm_data, after_scm=after_scm, after_recipe=after_recipe)
-
-    @pytest.mark.tool_git
-    def test_scm_from_superclass(self):
-        client = TurboTestClient()
-        conanfile = '''from conans import ConanFile
-
-def get_conanfile():
-
-    class BaseConanFile(ConanFile):
-        scm = {
-            "type": "git",
-            "url": "auto",
-            "revision": "auto"
-        }
-
-    return BaseConanFile
-
-class Baseline(ConanFile):
-    name = "Base"
-    version = "1.0.0"
-'''
-        client.init_git_repo({"conanfile.py": conanfile}, origin_url="http://whatever.com/c.git")
-        client.run("export . conan/stable")
-        conanfile1 = """from conans import ConanFile, python_requires, tools
-
-baseline = "Base/1.0.0@conan/stable"
-
-# recipe inherits properties from the conanfile defined in the baseline
-class ModuleConan(python_requires(baseline).get_conanfile()):
-    name = "module_name"
-    version = "1.0.0"
-"""
-        conanfile2 = """from conans import ConanFile, python_requires, tools
-
-baseline = "Base/1.0.0@conan/stable"
-
-# recipe inherits properties from the conanfile defined in the baseline
-class ModuleConan(python_requires(baseline).get_conanfile()):
-    pass
-"""
-
-        for conanfile in [conanfile1, conanfile2]:
-            client.save({"conanfile.py": conanfile})
-            # Add and commit so it do the scm replacements correctly
-            client.run_command("git add .")
-            client.run_command('git commit -m  "commiting"')
-            client.run("export . module_name/1.0.0@conan/stable")
-            self.assertIn("module_name/1.0.0@conan/stable: "
-                          "A new conanfile.py version was exported", client.out)
-            ref = ConanFileReference.loads("module_name/1.0.0@conan/stable")
-            contents = load(os.path.join(client.cache.package_layout(ref).export(),
-                                         "conanfile.py"))
-            class_str = 'class ModuleConan(python_requires(baseline).get_conanfile()):\n'
-            self.assertIn('%s    scm = {"revision":' % class_str, contents)
-
-
-class SCMUpload(unittest.TestCase):
-
-    @pytest.mark.tool_git
-    def test_scm_sources(self):
-        """ Test conan_sources.tgz is deleted in server when removing 'exports_sources' and using
-        'scm'"""
-        conanfile = """from conans import ConanFile
-class TestConan(ConanFile):
-    name = "test"
-    version = "1.0"
-"""
-        exports_sources = """
-    exports_sources = "include/*"
-"""
-        servers = {"upload_repo": TestServer([("*/*@*/*", "*")], [("*/*@*/*", "*")],
-                                             users={"lasote": "mypass"})}
-        client = TestClient(servers=servers, users={"upload_repo": [("lasote", "mypass")]})
-        client.save({"conanfile.py": conanfile + exports_sources, "include/file": "content"})
-        client.run("create . danimtb/testing")
-        client.run("upload test/1.0@danimtb/testing -r upload_repo")
-        self.assertIn("Uploading conan_sources.tgz", client.out)
-        ref = ConanFileReference("test", "1.0", "danimtb", "testing")
-        rev = servers["upload_repo"].server_store.get_last_revision(ref).revision
-        ref = ref.copy_with_rev(rev)
-        export_sources_path = os.path.join(servers["upload_repo"].server_store.export(ref),
-                                           "conan_sources.tgz")
-        self.assertTrue(os.path.exists(export_sources_path))
-
-        scm = """
-    scm = {"type": "git",
-           "url": "auto",
-           "revision": "auto"}
-"""
-        client.save({"conanfile.py": conanfile + scm})
-        client.run_command("git init")
-        client.run_command('git config user.email "you@example.com"')
-        client.run_command('git config user.name "Your Name"')
-        client.run_command("git remote add origin https://github.com/fake/fake.git")
-        client.run_command("git add .")
-        client.run_command("git commit -m \"initial commit\"")
-        client.run("create . danimtb/testing")
-        self.assertIn("Repo origin deduced by 'auto': https://github.com/fake/fake.git", client.out)
-        client.run("upload test/1.0@danimtb/testing -r upload_repo")
-        self.assertNotIn("Uploading conan_sources.tgz", client.out)
-        rev = servers["upload_repo"].server_store.get_last_revision(ref).revision
-        ref = ref.copy_with_rev(rev)
-        export_sources_path = os.path.join(servers["upload_repo"].server_store.export(ref),
-                                           "conan_sources.tgz")
-        self.assertFalse(os.path.exists(export_sources_path))
